Combined options and main
